+++ Rendering/vtkCuller.h
@@ -38,8 +38,8 @@ public:
 
   // Description:
   // This is called outside the render loop by vtkRenderer
-  virtual float Cull( vtkRenderer *ren, vtkProp **propList,
-                      int& listLength, int& initialized )=0;
+  virtual double Cull( vtkRenderer *ren, vtkProp **propList,
+                       int& listLength, int& initialized )=0;
 
 protected:
   vtkCuller();
